requirements.txt
python-telegram-bot>=22.5
pandas>=2.0.0
openpyxl>=3.1.0
requests>=2.31.0
python-dotenv>=1.0.0

